% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgPersist.r
\name{avgPersist}
\alias{avgPersist}
\title{Compute average carcass persistence}
\usage{
avgPersist(searches, dist = "exponential", shape, scale)
}
\arguments{
\item{searches}{A vector containing julian dates of all
carcass searches. Only the number of searches and difference
between min and max matter.  Average search interval is
computed and average persistence probability is computed
between 0 and average interval.  Hence, placement of searches
within \code{min(searches)} and \code{max(searches)} does
not matter.}

\item{dist}{Character string naming the survival distribution
to use. Valid values are "exponential" and "weibull".}

\item{shape}{Shape parameter of the survival distribution.
If \code{dist} == "exponential", this parameter is ignored.}

\item{scale}{Scale parameter of the survival distribution.
If \code{dist} == "exponential", this parameter is the mean,
which is 1/rate, where rate is the parameter of \code{pexp} in R.}
}
\value{
A single number.  The average probability of a carcass
surviving from 0 to the average search interval. i.e.,
average survival under the survival curve from 0 to
\code{diff(range(searches))/2}.
}
\description{
This computes average carcass persistence
during a set of searches, given a cummulative hazard distribution
}
\examples{
avgPersist(c(0,15,30),"exponential",scale=8.6251)
avgPersist(c(0,15,30),"weibull",0.57792,7.1683)


}
\seealso{
\code{\link{eoa}}
}
\author{
Trent McDonald
}
