% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGwithK.R
\name{calcGwithK}
\alias{calcGwithK}
\title{calcGwithK - Calculate eoa's g value}
\usage{
calcGwithK(searchDates, crShape, crScale, crDist, seef, k, a)
}
\arguments{
\item{searchDates}{A vector of search dates.  This can either be a character
vector containing representations of real dates (e.g.,
c('2014-03-01','2014-04-01', etc.)), a vector of \code{POSIXct}
objects representing real dates, or a vector of integers specifying days
after start of the study (e.g., c(0,30,60, etc.)).  If \code{searchDates}
is a vector of integers, the first element must be 0.  If \code{searchDates}
is a vector of characters, format of the dates must be "%Y-%m-%d"
(or YYYY-MM-DD).}

\item{crShape}{Shape parameter of the persistence distribution for
objects. If \code{crDist} == "exponential", this parameter is ignored.}

\item{crScale}{Scale parameter of the persistence distribution for
objects. If \code{crDist} == "exponential", this parameter is the mean,
which is 1/rate, where rate is the parameter of \code{pexp} in R.}

\item{crDist}{Character string naming the persistence distribution
to use. Valid values are "exponential", "weibull", "lognormal", and "loglogistic".
Anything else throws an error.}

\item{seef}{A scalar specifying probability of detection. This is probability
of detecting an object given it is present.}

\item{k}{A scalar specifying the proportion of objects remaining available
for detection
following the first search they became available.  E.g., \code{k} of the
objects missed on a search are available during the next search.}

\item{a}{A scalar specifying the proportion of the object's spatial distribution
searched, sometimes called area correction proportion.}
}
\value{
A scalar, the probability of object discovery, which is the product
of probability of an object being available and probability of detection
given presence.
}
\description{
Compute EoA g = Pr(discovery) given searcher efficiency,
carcass removal distribution, area correction, "k", and search dates.
}
\details{
This routine assumes the arrival function of objects on
the study area is uniform between \code{min(searchDates)} and
\code{max(searchDates)}.
}
\examples{
days <- c(0,7,14,21,28,58,88)
calcGwithK(days, NA, 21, "exponential", 0.7, 0.2, 0.5)

days <- c("2014-05-15","2014-06-15","2014-08-15")
calcGwithK(days, NA, 21, "exponential", 0.7, 0.2, 0.5)

startDay <- c("2014-05-15")
endDay <- c("2014-09-15")
searchInterval <- 7 # days
nSearches <- floor(as.numeric(as.POSIXct(endDay) - as.POSIXct(startDay))/searchInterval)
days <- as.POSIXct(startDay) + (0:nSearches)*searchInterval*(60*60*24)
calcGwithK(days, 0.57792, 7.16, "weibull", 0.7, 0.2, 0.5)

}
\author{
Trent McDonald, with help from Jared Studyvin,
Kristen Nasmen, and Paul Rabie.  Based largely on code in package \code{eoa}
by Dan Dalthorp.
}
