% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkIsAutocorrelated.r
\name{checkIsAutocorrelated}
\alias{checkIsAutocorrelated}
\title{checkIsAutocorrelated - Check whether MCMC chains are autocorrelated.}
\usage{
checkIsAutocorrelated(obj, criterion = 0.4, lag = 1, quiet = FALSE)
}
\arguments{
\item{obj}{An mcmc.list object from the \code{rjags} package.  These are output
by the \code{coda.samples} function.}

\item{criterion}{scalar specifying the level of acceptable autocorrelation at
the assessed \code{lag}.  If all average lag2 autocorrelations are less than this number, the
whole mcmc object is deems 'random' or not autocorrelated.}

\item{lag}{The lag at which to assess autocorrelation.  This should generally be a
low number like 1, 2, 5, or 10.}

\item{quiet}{Whether to print any output on screen.}
}
\value{
A list with the following components
\itemize{
  \item autoCorrelated: If the average autocorrelations at the specified
  lag are all less than
  \code{criterion}, the chain is declared un-autocorrelated and is set FALSE.
  Otherwise, this component of the returns is TRUE.
  \item autoCorrs: the computed autocorrelations at the specified lag.
  This value is the average across chains in \code{obj}.
 }
}
\description{
Run autocorrelation on every parameter in an mcmc.list and make
a decision whether they are significantly auto correlated.  Ideally, chains
are not autocorrelated when converged.
}
\details{
This routine calculates auto correlation separately at the specified lag
for every chain in the mcmc.list, then averages autocorrelations across
chains.
}
\examples{
\dontrun{
jags = jags.model(file="someJAGSFile.txt",
   data=someJAGS.data, inits=someJAGS.inits,
   n.chains=3,n.adapt=100)
update(jags, n.iter=1000)
out = coda.samples(jags,
   variable.names=c("someParameter"),
   n.iter=1000,
   thin=2)

auto <- chechIsAutocorrelated(out)
}
}
\seealso{
\code{\link{checkIsConverged}}
}
\author{
Trent McDonald
}
