% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkIsConverged.r
\name{checkIsConverged}
\alias{checkIsConverged}
\title{checkConvergence - Check the convergence of a mcmc.list object}
\usage{
checkIsConverged(obj, criterion = 1.1, quiet = FALSE)
}
\arguments{
\item{obj}{An mcmc.list object from the \code{rjags} package.  These are output
by the \code{coda.samples} function. This mcmc.list must contain 2 or more chains.}

\item{quiet}{Whether to print any output on screen.}
}
\value{
A list with the following components
\itemize{
  \item converged: If the uppper limits of the  'potential scale
  reduction factors' associated with all parameters are less than
  \code{criterion}, the chain is declared converged and this is TRUE.
  Otherwise, this component of the returns is FALSE.
  \item Rhats: the 'potential scale reduction factors' or Rhats
  associated with all parameters.
 }
}
\description{
Run various diagnostic procedures in the \code{coda} package
to check convergence of a set of mcmc chains.
}
\details{
This routine runs the Gelman and Rubin (1992) diagnostic for
a set of MCMC chains by calling \code{gelman.diag} in the \code{coda} package.
}
\examples{
\dontrun{
jags = jags.model(file="someJAGSFile.txt",
   data=someJAGS.data, inits=someJAGS.inits,
   n.chains=3,n.adapt=100)
update(jags, n.iter=1000)
out = coda.samples(jags,
   variable.names=c("someParameter"),
   n.iter=1000,
   thin=2)

conf <- chechIsConverged(out)
}
}
\references{
Gelman, A and Rubin, DB (1992) Inference from iterative simulation
using multiple sequences, Statistical Science, 7, 457-511.
}
\seealso{
\code{\link{gelman.diag}}, \code{\link{checkIsConverged}}
}
\author{
Trent McDonald
}
