% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compVagueSd.r
\name{compVagueSd}
\alias{compVagueSd}
\title{compVagueSd - Compute sD's of vague priors}
\usage{
compVagueSd(Y, alpha.vec, beta.vec, X, range.multiplier = 100)
}
\arguments{
\item{Y}{Vector of number of carcasses found, one element per year. If
multiple sites are involved, elements of Y are the total (summed) number
of targets found per season.}

\item{alpha.vec}{vector of the alpha parameters of the Beta distributions}

\item{beta.vec}{vector of the beta parameters of the Beta distributions}

\item{X}{a design matrix upon which an approximation of inflated
numbers of targets is regressed.  Usually, this is meant to be the
design matrix from the \code{eoa} function.}

\item{range.multiplier}{a multipiler for the range of coefficient
estimates to make the output standard deviations sufficiently vague.
Increasing this number increases vagueness.}
}
\value{
List containing two components.  \code{$vagueSd} is a
vector, one per parameter, of standard deviations that should
be large enough to call vague when used in a normal prior.
\code{$startA} is a vector of potential starting values for
the coefficients in the model.
}
\description{
Compute the standard deviation of a normal
distribution that is big enough to be considered a 'vague'
prior.  This is not straight forward when there are covariates, as here.
}
