% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateM.EoA.R
\name{estimateM.EoA}
\alias{estimateM.EoA}
\title{estimateM.EoA}
\usage{
estimateM.EoA(X, beta.params, Mprior = "objective", Mprior.mean, Mprior.sd,
  conf.level = 0.9)
}
\arguments{
\item{X}{Total number of carcasses found.}

\item{beta.params}{A list or data frame containing at a minimum
components named \code{$alpha} and \code{$beta}.
These are the alpha and beta parameters of a Beta distribution which
is used for g=Pr(discovery).}

\item{Mprior}{Character string specifying the prior distribution
for M.
 \itemize{
   \item "objective" uses an objective prior very close to the
   Jeffery's prior for a Poisson, i.e., sqrt(m+1)-sqrt(m).
   \item "normal" uses a truncated and descretized normal(Mprior.mean,Mprior.sd).
   \item "gamma" uses a descretized gamma with mean Mprior.mean and
   standard deviation Mprior.sd.  Note, this always assigns zero prior probability
   to M=0.
 }}

\item{Mprior.mean}{Mean of M prior when Mprior == "normal" or "gamma".}

\item{Mprior.sd}{Standard deviation of M prior when Mprior == "normal" or "gamma".}

\item{conf.level}{Confidence level for the confidence intervals on
posterior estimates of M and g.}
}
\value{
List containing the following components.
\itemize{
  \item \code{M.est} : A data
   frame containing the following:
   \itemize{
     \item \code{M} = usual point estimate of M = median of M posterior
     distribution.
     \item \code{M.mu} = mean of M posterior distribution
     \item \code{M.sd} = standard deviation of M posterior distribution
     \item \code{M.lo} = lower endpoint of a 100(conf.level)% posterior credible
     interval for M.
     \item \code{M.hi} = upper endpoint of a 100(conf.level)% posterior credible
     interval for M.
     \item \code{g} = mean of g posterior distribution
     \item \code{g.lo} = lower endpoint of a 100(conf.level)% posterior credible
     interval for g.  Note, this does not agree with the analogous number
     from package \code{eoa} because this comes from the posterior for g.
     \code{eoa} reports the quantile from the prior for g.
     \item \code{g.hi} = upper endpoint of a 100(conf.level)% posterior credible
     interval for g. Note, this does not agree with the analogous number
     from package \code{eoa} because this comes from the posterior for g.
     \code{eoa} reports the quantile from the prior for g.
     \item \code{ci.level} = confidence level of the credible intervals (same
     as input \code{conf.level}).
   }


  \item \code{M.margin} : the full posterior marginal distribution for M.
  This is a data frame with the following columns
  \itemize{

    \item \code{M} : value of M in its support.
    \item \code{pdf} : posterior probability mass function for M.  Pr(M=m)
    \item \code{cdf} : posterior cummulative probability mass function for M.
    Pr(M<=m).
    \item \code{prior.pdf} : prior probability mass function for M.
    \item \code{like.pdf} : likelihood probability mass function for M.
  }
  Note, all three of the pdf columns sum to 1.0, even in the
  case of an improper prior. These columns can be plotted together
  using the plot method for \code{Mest} objects.

  \item \code{g.margin} : the full posterior marginal distribution for g.
  This is a data frame with columns \code{$g}, \code{$pdf.g}, and \code{$cdf.g}
  corresponding to g, probability of g, and probability of being less than or
  equal to g, respectively. Note, \code{sum(result$g.margin$pdf.g)}
  \code{*diff(result$g.margin$g)[1] == 1.0}.
 }
}
\description{
Estimate single-site or multiple-class M (=mortalities)
parameter of Evidence of Absence (EoA)
using objective or informed priors.
}
\details{
This routine replicates the M estimates of the 'Single Year' and
'Multiple Classes' modules in package \code{eoa}.  To repeat either case,
input the composite g parameter's "a" and "b" parameters here, along
with the number of carcasses "X", and specify the "objective" prior. See
Examples.
}
\examples{
g.params <- list(alpha=600, beta=1200)
X <- 5
m.ests <- estimateM.EoA(X,g.params)
print(m.ests$M.est)

m.ests <- estimateM.EoA(X, g.params, Mprior = "normal", Mprior.mean = 50, Mprior.sd = 30)
print(m.ests$M.est)

m.ests <- estimateM.EoA(X, g.params, Mprior = "gamma", Mprior.mean = 50, Mprior.sd = 30)
print(m.ests$M.est)
}
\seealso{
\code{\link{estimateL.EoA}}, \code{\link{plot.Mest}}
}
\author{
Trent McDonald
}
