% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmmax.ab.R
\name{fmmax.ab}
\alias{fmmax.ab}
\title{fmmax.ab}
\usage{
fmmax.ab(x, pBa, pBb)
}
\arguments{
\item{x}{The number of carcasses (number of successes).}

\item{pBa}{The alpha parameter of g's Beta distribution}

\item{pBb}{The beta parameter of g's Beta distribution}
}
\value{
The M with less than 0.0001 probability of occurring.
}
\description{
Compute the maximum M (mortality) we need to worry about
during EoA and EoAR calculations.
}
\examples{
beta.params <- data.frame(alpha=231.59, beta=2239.925)
x <- 40
fmmax.ab(x, beta.params$alpha, beta.params$beta)

}
\author{
Dan Dalthorp
}
