% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFleetG.R
\name{getFleetG}
\alias{getFleetG}
\title{getFleetG - Compute all-site (fleetwide) Evidence of Absence g-value.}
\usage{
getFleetG(siteYearResults, species = "LBBA", weights = NULL)
}
\arguments{
\item{siteYearResults}{A data frame containing site-season results.
One usually obtains this by running the site-year module of the \code{eoa}
package on
every combination of year and site.  To assist with this, we at WEST have a python app
which copies and pastes the input parameters from CSV files into the
interface of
\code{eoa}.
Once site-year g-values are computed (using the auto-copy-paste routine),
we have a "scraper" function in R which
pulls parameters from all the text files that \code{eoa} produces.
This parameter is a data frame containing the output CSV results
from the "scraper".
At a minimum, \code{siteYearResults} must contain the following
columns:  \code{$species}, \code{$facility}, \code{$gFac.a} = the
alpha paramters of this facility's beta distribution, \code{$gFac.b} =
the beta parameter of this facility's beta distribution, and \code{$year}.}

\item{species}{Species abbreviation to run.}

\item{weights}{A scalar or data frame containing weights for each
facility in the fleet. One row per
facility.  This is the "DWP" field of \code{eoa}. If \code{weights} = NULL,
all facilities receive equal weight.  Otherwise, \code{weights} must be
a data frame containing \code{$facility} and \code{$weight} columns.  This is merged
with \code{siteYearResults} using \code{$facility} as the key.}
}
\value{
The fleet-wide g-value for a particular species.  This
value can then go into \code{estimateL.eoa}.
}
\description{
Compute a single all-site g value from site-season specifice g values.
}
\details{
Seasons are summarized first.  That is, if a site appears more than
once for a species, we assume each line is a separate season
for that species.  g is averaged over those lines first.  At the end,
this average goes into the multi-site g average.
}
\examples{
syr <- data.frame(species=c("LBBA","LBBA","LBBA"),
   facility=c("f1","f2","f2"),
   gFac.a = c( 69.9299, 63.5035,  84.6997),
   gFac.b = c(  736.4795,  318.3179, 759.9333 ),
   year = c(2015,2015,2016))
getFleetG(syr, "LBBA")
}

