% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.matrix.eoa.r
\name{model.matrix.eoa}
\alias{model.matrix.eoa}
\title{model.matrix.eoa}
\usage{
\method{model.matrix}{eoa}(object, ...)
}
\arguments{
\item{object}{An object of class \code{eoa}. In particular,
this object must have a $terms component.}

\item{\dots}{Included for compatability with other methods}
}
\value{
The design matrix for an eoa object with all the factor
levels and interactions expanded.
}
\description{
Model matrix method for eoa objects produced
by the \code{eoa} function.
}
\author{
Trent McDonald
}
