% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Mest.r
\name{plot.Mest}
\alias{plot.Mest}
\title{Plot M posterior distribution}
\usage{
\method{plot}{Mest}(obj, plot.like = FALSE, plot.prior = FALSE)
}
\arguments{
\item{obj}{An object of class "Mest".  Usually output by the
routine \code{estimateM.EoA}.}
}
\description{
Plot method for the M estimates
(from \code{evoab}) posterior distribution.
}
