% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.eoa.r
\name{print.eoa}
\alias{print.eoa}
\title{Print a Evidence of Absence model}
\usage{
\method{print}{eoa}(x, ...)
}
\arguments{
\item{x}{An estimated eoa object from \code{eoa}.}

\item{\dots}{Included for compatibility with other print methods.  Ignored here.}
}
\value{
The input value of \code{obj} is invisibly returned.
}
\description{
Print method for EoA models produced by \code{eoa},
which are of class \code{eoa}.
}
\examples{

}
\seealso{
\code{\link{summary.eoa}}
}
\author{
Trent McDonald, WEST Inc. \email{tmcdonald@west-inc.com}
}
\keyword{models}
