% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.eoa.r
\name{update.eoa}
\alias{update.eoa}
\title{update.eoa - Update method for EoA objects}
\usage{
\method{update}{eoa}(obj, niters, nburns = 0, nthins = 1,
  add = FALSE, quiet = FALSE)
}
\arguments{
\item{obj}{An object of class \code{eoa}.  See function \code{\link{eoa}}.}

\item{niters}{Number of additional iterations to do.}

\item{nburns}{Number of additional iterations to burn.}

\item{nthins}{Thinning interval for additional iterations}

\item{add}{If TRUE, the additional \code{niters/nthin} kept iterations
will be appended to the previous iterations. If FALSE, the additional
\code{niters/nthin} kept iterations will overwrite the previous iterations,
effectively increasing the burn-in period of the MCMC model.}
}
\value{
An \code{eoa} model object.  See \code{\link{eoa}}
}
\description{
Update (run more iterations of) an EoA model object output by \code{eoa()}.
}
\examples{
# A 3 year study of 7 sites. 21 "cells". lambda change = 20/year
set.seed(9430834) # fixes Y and g of this example, but not the RNG's used in chains
ns <- 3

ny <- 7
g <- data.frame(
 alpha = rnorm(ns*ny,70,2),
 beta = rnorm(ns*ny,700,25)
)
Y <- rbinom(ns*ny, c(rep(20,ny), rep(40,ny), rep(60,ny)), g$alpha/(g$alpha+g$beta))

df <- data.frame(year=factor(c(rep("2015",ny),rep("2016",ny),rep("2017",ny))),
   Year=c(rep(1,ny),rep(2,ny),rep(3,ny)))

# Uninformed eoa (use low number of iterations because it's and example)
eoa.1 <- eoa(Y~year, g, df, nburn = 1000, niters= 50*10, nthins = 10 )

# Could do this if eoa.1 did not converge
#  1) Consider everything in eoa.1 burn-in. (nadapt+nburn+niters iterations)
#  2) Burn another 1000 iterations, then sample 1000 taking every 10th
eoa.2 <- update(eoa.1, nburns=1000, niters=1000, nthin=10)

# Could do the following if eoa.1 converged, but autocorrelation was high.
#  1) Consider everything in eoa.1 burn-in. (nadapt+nburn+niters iterations)
#  2) Do another 1000 iterations with higher thinning
eoa.2 <- update(eoa.1, nburns=0, niters=1000, nthin=50)

# Could do the following if eoa.1 converged and autocorrelation were low,
# but simulation error was high (i.e., need more iterations)
#  1) Consider the n.iters in eoa.1 "good".
#  2) Add another 1000 iterations with same (or different) thinning
eoa.2 <- update(eoa.1, nburns=0, niters=1000, nthin=50, add=TRUE)

}
\seealso{
\code{\link{eoa}}, \code{\link{coef.eoa}}.
}
\author{
Trent McDonald
}
