% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Mest.r
\name{plot.Mest}
\alias{plot.Mest}
\title{Plot M posterior distribution}
\usage{
\method{plot}{Mest}(obj)
}
\arguments{
\item{obj}{An object of class "Mest".  Usually output by the
routine \code{estimateM.EoA}.}
}
\description{
Plot the M (from \code{evoab}) posterior distribution.
}
