% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.eoa.r
\name{predict.eoa}
\alias{predict.eoa}
\title{predict.eoa}
\usage{
\method{predict}{eoa}(object, newdata, ...)
}
\arguments{
\item{object}{an eoa object produced by function \code{eoa}}

\item{newdata}{A data frame containing new covariate
values at which to do the prediction.}

\item{\dots}{Included for compatability with other predict methods.}
}
\value{
If \code{newdata} is NULL, the predicted
lambda values for every row of the fitting data set.
If \code{newdata} is not NULL, results of applying
the lambda model to covarates in \code{newdata}.
}
\description{
Prediction method for eoa objects.
}
\author{
Trent McDonald
}
